+++ test/string.test.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { lower, sub, upper, words } from '../src/string.js';
+import { lower, sub, upper, validate, words } from '../src/string.js';
 
 describe('lower', () => {
   it('lower:text', () => {
@@ -24,6 +24,38 @@ describe('sub', () => {
 describe('upper', () => {
   it('upper:text', () => {
     expect(upper('lorem')).toEqual('LOREM');
+  });
+});
+
+describe('validate', () => {
+  it('validate:1', () => {
+    expect(
+      validate('HelloWorld123!', {
+        minimum: 8,
+        maximum: 15,
+        lowercase: 2,
+        uppercase: 2,
+        special: 1,
+        number: 2,
+        require: ['Hello'],
+        disable: ['World'],
+      })
+    ).toEqual([
+      'minimum',
+      'maximum',
+      'lowercase',
+      'uppercase',
+      'special',
+      'number',
+      'require',
+      'disable',
+    ]);
+  });
+
+  it('validate:2', () => {
+    expect(
+      validate('Password123!', { minimum: 8, lowercase: 2, uppercase: 2, special: 2, number: 2 })
+    ).toEqual(['minimum', 'maximum', 'lowercase', 'number']);
   });
 });